Classwork/requirements.txt
python-dotenv>=1.0.1
openai>=1.68.2
google-genai>=1.33.0
gradio>=5.22.0
groq>=0.31.1
pypdf>=3.14.0
crewai>=0.9.3
crewai-tools>=0.9.3
langgraph>=0.3.3
langchain_openai>=0.0.10
langchain_groq >=0.0.6